2004-05-09 23:33 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/include/hbdefs.h     * new macros HB_{GET|PUT}_{LE|BE}_UINT{24|64} to retrieve/store       24 and 64 bit integers in endian independent way
  * xharbour/source/rdd/rddord.prg
    + ORDWILDSEEK(cPatern, [lCont]) added (rather to illustrate posibilities)
      of new DBOI_SKIP{WILD|EVAL|REGEX} then final function. After beta1
      release I'll implement all the ORD{WILD|EVAL|REGEX}SEEK functions on
      C level

  * xharbour/source/rdd/dbcmd.c
  * xharbour/source/rdd/dbf1.c
  * xharbour/source/rdd/workarea.c
  * xharbour/source/rdd/dbffpt/dbffpt1.c
  * xharbour/source/rdd/dbfdpt/dbfdbt1.c
    + added suport in DBF files for: "I" (binary integer, len: 2, 4, 8),
      "D" (binary date, 3), "B" (binary double, 8) fields and some small
      fixes for "M" (4)

  * xharbour/source/rdd/dbfntx/dbfntx1.c
    ! fixed problem with SOFTSEEK on on filtered database int the same
      way as Alexander done in Harbour

  * xharbour/source/rdd/dbfcdx/dbfcdx1.c
    * casting added

  * xharbour/source/rtl/gtxvt/gtxvt.c
    ! removed unexisting macros HB_INKEY_RAW and HB_INKEY_EXTENDED

+++ xharbour/source/rdd/dbfcdx/dbfcdx1.c
@@ -1,5 +1,5 @@
 /*
- * $Id: dbfcdx1.c,v 1.129 2004/05/02 13:44:05 druzus Exp $
+ * $Id: dbfcdx1.c,v 1.130 2004/05/03 17:29:20 druzus Exp $
  */
 
 /*
@@ -624,7 +624,7 @@ static LPCDXKEY hb_cdxKeyPutItem( LPCDXKEY pKey, PHB_ITEM pItem, ULONG ulRec, LP
    {
       case HB_IT_STRING:
       case HB_IT_STRING | HB_IT_MEMO:
-         len = ( BYTE ) HB_MIN( pItem->item.asString.length, pTag->uiLen );
+         len = (BYTE) HB_MIN( pItem->item.asString.length, (ULONG) pTag->uiLen );
          if ( fSize && len < pTag->uiLen )
          {
             memcpy( ptr, pItem->item.asString.value, len );
